+++ lib/danger/helpers/comment.rb
@@ -2,9 +2,10 @@ module Danger
   class Comment
     attr_reader :id, :body
 
-    def initialize(id, body)
+    def initialize(id, body, inline = nil)
       @id = id
       @body = body
+      @inline = inline
     end
 
     def self.from_github(comment)
@@ -13,9 +14,10 @@ module Danger
 
     def self.from_gitlab(comment)
       if comment.respond_to?(:id) && comment.respond_to?(:body)
-        self.new(comment.id, comment.body)
+        type = comment.respond_to?(:type) ? comment.type : nil
+        self.new(comment.id, comment.body, type == "DiffNote")
       else
-        self.new(comment["id"], comment["body"])
+        self.new(comment["id"], comment["body"], comment["type"] == "DiffNote")
       end
     end
 
@@ -24,7 +26,7 @@ module Danger
     end
 
     def inline?
-      body.include?("")
+      @inline.nil? ? body.include?("") : @inline
     end
   end
 end